requirements.txt
streamlit>=1.37.0
pdfplumber>=0.11.0
pandas>=2.1.0
python-dateutil>=2.9.0
numpy>=1.26.0
